sbpt/pyproject.toml
[project]
name = "sbpt"
version = "0.1.0"
requires-python = ">=3.10,<3.11"
dependencies = [
  "torch>=2.3",
  "transformers>=4.44",
  "datasets>=2.20",
  "accelerate>=0.33",
  "numpy>=1.26,<2",
  "pydantic>=2.8",
  "rich>=13.7",
  "pyyaml>=6.0",
  "faiss-gpu==1.7.2",
]


[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
dev-dependencies = [
  "pytest>=8.2",
  "ruff>=0.5",
]
